toolkit/content/tests/chrome/test_textbox_emptytext.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<!--
  XUL Widget Test for textbox with placeholder
  -->
<window title="Textbox with placeholder test" width="500" height="600"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <hbox>
    <textbox id="t1"/>
  </hbox>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml" style="height: 300px; overflow: auto;"/>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

SimpleTest.waitForExplicitFinish();

function doTests() {
  var t1 = $("t1");

  t1.placeholder = 1;
  ok("1" === t1.label,          "placeholder exposed as label");
  ok(""  === t1.value,          "placeholder not exposed as value");

  t1.label = 2;
  ok("2" === t1.label,          "label can be set explicitly");
  ok("1" === t1.placeholder,    "placeholder persists after setting label");

  t1.value = 3;
  ok("3" === t1.value,          "value setter/getter works while placeholder is present");
  ok("1" === t1.placeholder,    "placeholder persists after setting value");

  t1.value = "";
  is(t1.textLength, 0,          "textLength while placeholder is displayed");

  SimpleTest.finish();
}

SimpleTest.waitForFocus(doTests);

  ]]></script>

</window>
